add moderateur app + create & connect local mysql database
