+++ core/Base_TP_IGL_MoteurRecherche.sql
@@ -24,4 +24,4 @@
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2023-12-13 16:03:08
+-- Dump completed on 2023-12-13 20:06:44
